==> 01_cleanup.ipynb <==
# Clean the original dataset and prepare it for further analyses

import numpy as np
import pandas as pd
import re
import utils
import swifter # swifter makes pandas.DataFrame.apply() slightly faster and communicates the progress

df_orig = pd.read_csv("xandr_segments.csv", sep=",")  # load data
# rename columns
df_orig = df_orig.rename(columns={"Segment Name": "name",
                                  "Data Provider Name": "provider_name",
                                  "Data Provider ID": "provider_id",
                                  "Segment ID": "id"})
df_orig["name_list"] = np.nan  # empty column for itemized segment name

# Providers by frequency
df_orig["provider_name"].value_counts()[:20]

# Clean segement names (diactrics, stripping, capitalization)
df_orig["name"] = df_orig.swifter.apply(lambda row: utils.clean_segment_name(row["name"]), axis = 1)

# Segments by a few brokers are multilingual which messes up all kinds of upcoming analyses. All of them follow the same scheme
# and are simply chopped off at this stage.
pruning_re = re.compile(r"\(en\).+\(es\).+\(pt\)")
def prune_segname(row):
    name = row["name"]
    if match := re.search(pruning_re, name):
         row["name"] = name[:match.start(0)].strip()
    return row

df_orig = df_orig.swifter.apply(prune_segname, axis=1)


# itemize all segment names
df_orig["name_list"] = df_orig.swifter.apply(lambda row: utils.itemize_segment_name(row["name"]), axis = 1)
df_orig = df_orig.set_index("id")

# remove segments without useful content
useless_segment_indices = df_orig.swifter.apply(lambda x: bool(re.search(utils.useless_segments_re, x["name"])), axis = 1)

df_useless_segments = df_orig[useless_segment_indices].reset_index(drop=True)
df_useless_segments.to_csv("xandr_segments_useless.csv")

df_filtered = df_orig[useless_segment_indices.__invert__()]
df_filtered.to_csv('xandr_segments_itemized.csv')

# sanity check for segment names that could not be itemized
df_failed_itemization = df_filtered[df_filtered.swifter.apply(lambda x: len(x["name_list"]) < 2, axis = 1)]
df_failed_itemization = df_failed_itemization.reset_index(drop=True)
df_failed_itemization.to_csv('xandr_segments_failed_itemization.csv')
